Minha mensagem do commit

+++ src/domain/Price.java
@@ -6,18 +6,18 @@ import java.util.UUID;
 
 public class Price implements EntityInterface {
     private UUID uuid;
-    private BigDecimal price;
+    private Float price;
     private Date date;
 
     public Price() {
     }
 
-    public Price(BigDecimal price, Date date) {
+    public Price(Float price, Date date) {
         this.price = price;
         this.date = date;
     }
 
-    public Price(UUID uuid, BigDecimal price, Date date) {
+    public Price(UUID uuid, Float price, Date date) {
         this.uuid = uuid;
         this.price = price;
         this.date = date;
